Создаст 'MovieCardFull' и 'MovieCardPromo'.
