+++ src/components/movie-card-full/movie-card.test.js
